+++ src/test/java/com/cisowski/schoolmanagement/integration/helper/TestDataHelper.java
@@ -163,6 +163,19 @@ public class TestDataHelper {
         return scheduleVersionRepository.save(scheduleVersion);
     }
 
+    public ScheduleVersionEntity createScheduleVersion(YearbookEntity yearbook, boolean isActive) {
+        if(yearbook == null)
+            yearbook = createYearbook(Collections.emptyList(), null);
+        ScheduleVersionEntity scheduleVersion = Instancio.of(ScheduleVersionEntity.class)
+                .set(field(ScheduleVersionEntity::getId), null)
+                .set(field(ScheduleVersionEntity::getYearbook), yearbook)
+                .set(field(ScheduleVersionEntity::getSchedules), null)
+                .set(field(ScheduleVersionEntity::getStatus), ScheduleStatus.SCHEDULED)
+                .set(field(ScheduleVersionEntity::isActive), isActive)
+                .create();
+        return scheduleVersionRepository.save(scheduleVersion);
+    }
+
     public YearbookEntity createYearbook(List<SubjectEntity> subjects, TeacherEntity teacher) {
         YearbookEntity yearbook = Instancio.of(YearbookEntity.class)
                 .set(field(YearbookEntity::getHeadTeacher), teacher)