+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 ### New
 - Add refresh JWT token endpoint - restricted to authenticated users
 - Add 'me' endpoint returning base User data
+- Add missing endpoint to retrieve active ScheduleVersion for Student's Yearbook
 
 ### Fixed
 - Fix DB support for polish characters